Обновлены модели магазинов
