+++ DiscountsAndPromotionsApp/Network/ProductNetworkServices/Models/LocationResponseModel.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct LocationResponseModel: Codable {
+   let id: Int
+   let region, city, street, building: String?
+}
