+++ DiscountsAndPromotionsApp/Network/ProductNetworkServices/Models/StoreChainResponseModel.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct StoreChainResponseModel: Codable {
+    let id: Int
+    let name: String?
+}
+
+typealias StoreChainsResponseModel = [StoreChainResponseModel]
